feat: Add LoginPage and update main.dart to use it
This commit adds a new file called login.dart which contains the LoginPage widget. The LoginPage is a StatelessWidget that displays a simple login page with a centered text widget showing "Login Page" as the title. The main.dart file is updated to import and use the LoginPage instead of the MyHomePage, redirecting the user to the login page instead of the home page when the app is launched. This change improves the user experience by setting the foundation for implementing the login functionality in future commits.