feat: Add functionality to add pizzas to cart through API
Enhance the MyHomePage component by adding functionality to add pizzas to the cart through the API. The _addToCart method is updated to send a POST request to the '/cart/post.php' endpoint with the user ID and pizza ID. The response is handled to show appropriate feedback to the user. This change improves the application by providing a convenient way to add pizzas to the cart dynamically.