feat: Add functionality to update order status in admin_orders.dart
This commit adds the `updateOrderStatus` method to the `_AdminOrdersPageState` class in `admin_orders.dart`. The method sends a POST request to the `/order/update_status.php` API endpoint with the order ID and the new status. If the update is successful, a success message is shown and the order list is refreshed. If there is an error, an error message is shown. This change enhances the functionality of the application by allowing administrators to update the status of orders through the UI.