feat: Delete cart items after creating an order
This commit adds code to delete the cart items after successfully creating an order in the `post.php` file. It removes the cart items associated with the user by sending a DELETE query to the database. If the deletion is unsuccessful, an exception is thrown. This change enhances the functionality of the application by ensuring that cart items are properly cleared after an order is placed.