feat: Add registration link to login page
This commit adds a registration link to the login page. Users who do not have an account can now click on the link to navigate to the registration page and create a new account. This change enhances the user experience by providing a convenient way for new users to sign up for the application.